[src/platform] Separation of nrfconnect and Zephyr platforms
nrfconnect platform is based on the Zephyr and in many cases it uses
Zephyr specific components implementation. Zephyr specific files should
be located in the Zephyr platform.

* Moved Zephyr specific modules from the nrfconnect to the Zephyr platform
* Moved BLEManagerImpl and ThreadStackManagerImpl implementations from Generic_Zephyr files.
* Removed Generic_Zephyr files.
* Aligned READMEs for nrfconnect and Zephyr platforms.
