+++ src/include/platform/ThreadStackManager.h
@@ -35,8 +35,6 @@ class DeviceNetworkInfo;
 class DeviceControlServer;
 class BLEManagerImpl;
 template <class>
-class GenericBLEManagerImpl_Zephyr;
-template <class>
 class GenericPlatformManagerImpl;
 template <class>
 class GenericConfigurationManagerImpl;
@@ -89,8 +87,6 @@ private:
     friend class ConfigurationManagerImpl;
 #if CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
     friend class Internal::BLEManagerImpl;
-    template <class>
-    friend class Internal::GenericBLEManagerImpl_Zephyr;
 #endif
     friend class Internal::DeviceControlServer;
     template <class>
