--- src/platform/nrfconnect/BLEManagerImpl.cpp
@@ -1,40 +1,0 @@
-/*
- *
- *    Copyright (c) 2020 Project CHIP Authors
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-/**
- *    @file
- *          Provides an implementation of the BLEManager singleton object
- *          for Zephyr platforms.
- */
-
-#include <platform/internal/CHIPDeviceLayerInternal.h>
-
-#if CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
-
-#include <platform/Zephyr/GenericBLEManagerImpl_Zephyr.cpp>
-
-namespace chip {
-namespace DeviceLayer {
-namespace Internal {
-
-BLEManagerImpl BLEManagerImpl::sInstance;
-
-} // namespace Internal
-} // namespace DeviceLayer
-} // namespace chip
-
-#endif // CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
