+++ src/platform/nrfconnect/BLEManagerImpl.h
@@ -17,68 +17,11 @@
 
 /**
  *    @file
- *          Provides an implementation of the BLEManager singleton object
- *          for the Zephyr platforms.
+ *          Provides an implementation of the BLEManager object
+ *          for nRF Connect SDK platforms, by including Zephyr platform
+ *          implementation.
  */
 
 #pragma once
 
-#if CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
-
-#include <platform/Zephyr/GenericBLEManagerImpl_Zephyr.h>
-
-namespace chip {
-namespace DeviceLayer {
-namespace Internal {
-
-using namespace chip::Ble;
-
-/**
- * Concrete implementation of the NetworkProvisioningServer singleton object for the Zephyr platforms.
- */
-class BLEManagerImpl final : public GenericBLEManagerImpl_Zephyr<BLEManagerImpl>
-{
-    // Allow the BLEManager interface class to delegate method calls to
-    // the implementation methods provided by this class.
-    friend BLEManager;
-
-private:
-    // ===== Members that implement the GenericBLEManagerImpl_Zephyr internal interface.
-
-    // None so far.
-
-    // ===== Members for internal use by the following friends.
-
-    friend BLEManager & BLEMgr(void);
-    friend BLEManagerImpl & BLEMgrImpl(void);
-
-    static BLEManagerImpl sInstance;
-};
-
-/**
- * Returns a reference to the public interface of the BLEManager singleton object.
- *
- * Internal components should use this to access features of the BLEManager object
- * that are common to all platforms.
- */
-inline BLEManager & BLEMgr(void)
-{
-    return BLEManagerImpl::sInstance;
-}
-
-/**
- * Returns the platform-specific implementation of the BLEManager singleton object.
- *
- * Internal components can use this to gain access to features of the BLEManager
- * that are specific to the Zephyr platforms.
- */
-inline BLEManagerImpl & BLEMgrImpl(void)
-{
-    return BLEManagerImpl::sInstance;
-}
-
-} // namespace Internal
-} // namespace DeviceLayer
-} // namespace chip
-
-#endif // CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
+#include <platform/Zephyr/BLEManagerImpl.h>
