+++ src/platform/nrfconnect/ConfigurationManagerImpl.h
@@ -18,101 +18,10 @@
 /**
  *    @file
  *          Provides an implementation of the ConfigurationManager object
- *          for Zephyr platforms.
+ *          for nRF Connect SDK platforms, by including Zephyr platform
+ *          implementation.
  */
 
 #pragma once
 
-#include <platform/internal/GenericConfigurationManagerImpl.h>
-
-#include <platform/nrfconnect/ZephyrConfig.h>
-
-namespace chip {
-namespace DeviceLayer {
-
-/**
- * Concrete implementation of the ConfigurationManager singleton object for the Zephyr platform.
- */
-class ConfigurationManagerImpl final : public ConfigurationManager,
-                                       public Internal::GenericConfigurationManagerImpl<ConfigurationManagerImpl>,
-                                       private Internal::ZephyrConfig
-{
-    // Allow the ConfigurationManager interface class to delegate method calls to
-    // the implementation methods provided by this class.
-    friend class ConfigurationManager;
-
-    // Allow the GenericConfigurationManagerImpl base class to access helper methods and types
-    // defined on this class.
-#ifndef DOXYGEN_SHOULD_SKIP_THIS
-    friend class Internal::GenericConfigurationManagerImpl<ConfigurationManagerImpl>;
-#endif
-
-private:
-    // ===== Members that implement the ConfigurationManager public interface.
-
-    CHIP_ERROR _Init(void);
-    CHIP_ERROR _GetPrimaryWiFiMACAddress(uint8_t * buf);
-    bool _CanFactoryReset(void);
-    void _InitiateFactoryReset(void);
-    CHIP_ERROR _ReadPersistedStorageValue(::chip::Platform::PersistedStorage::Key key, uint32_t & value);
-    CHIP_ERROR _WritePersistedStorageValue(::chip::Platform::PersistedStorage::Key key, uint32_t value);
-
-    // NOTE: Other public interface methods are implemented by GenericConfigurationManagerImpl<>.
-
-    // ===== Members for internal use by the following friends.
-
-    friend ConfigurationManager & ConfigurationMgr(void);
-    friend ConfigurationManagerImpl & ConfigurationMgrImpl(void);
-
-    static ConfigurationManagerImpl sInstance;
-
-    // ===== Private members reserved for use by this class only.
-
-    static void DoFactoryReset(intptr_t arg);
-};
-
-/**
- * Returns the public interface of the ConfigurationManager singleton object.
- *
- * chip applications should use this to access features of the ConfigurationManager object
- * that are common to all platforms.
- */
-inline ConfigurationManager & ConfigurationMgr(void)
-{
-    return ConfigurationManagerImpl::sInstance;
-}
-
-/**
- * Returns the platform-specific implementation of the ConfigurationManager singleton object.
- *
- * chip applications can use this to gain access to features of the ConfigurationManager
- * that are specific to the nRF Connect SDK platform.
- */
-inline ConfigurationManagerImpl & ConfigurationMgrImpl(void)
-{
-    return ConfigurationManagerImpl::sInstance;
-}
-
-inline bool ConfigurationManagerImpl::_CanFactoryReset()
-{
-    return true;
-}
-
-inline CHIP_ERROR ConfigurationManagerImpl::_ReadPersistedStorageValue(::chip::Platform::PersistedStorage::Key key,
-                                                                       uint32_t & value)
-{
-    return ReadConfigValueCounter(key, value);
-}
-
-inline CHIP_ERROR ConfigurationManagerImpl::_WritePersistedStorageValue(::chip::Platform::PersistedStorage::Key key, uint32_t value)
-{
-    return WriteConfigValueCounter(key, value);
-}
-
-inline CHIP_ERROR ConfigurationManagerImpl::_GetPrimaryWiFiMACAddress(uint8_t * /* buf */)
-{
-    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
-}
-
-} // namespace DeviceLayer
-} // namespace chip
+#include <platform/Zephyr/ConfigurationManagerImpl.h>
