+++ src/platform/nrfconnect/README.md
@@ -5,22 +5,13 @@ this code will have parallels in any new adaptation.
 
 (All file names are relative to `connectedhomeip/src/...`).
 
-`include/platform/nrfconnect/PlatformManagerImpl.h`<br>`nrfconnect/PlatformManagerImpl.cpp`
+`include/platform/nrfconnect/PlatformManagerImpl.h`
 
--   Concrete implementation of PlatformManager interface
--   Provides initialization of the CHIP stack and core event loop for the chip
-    task
--   Relies on GenericPlatformManagerImpl_POSIX<> class to provide most of the
-    implementation
+-   Uses Zephyr platform implementation of PlatformManager interface.
 
-`include/platform/nrfconnect/ConfigurationManagerImpl.h`<br>`nrfconnect/ConfigurationManagerImpl.cpp`
+`include/platform/nrfconnect/ConfigurationManagerImpl.h`
 
--   Concrete implementation of ConfigurationManager interface
--   Manages storage and retrieval of persistent configuration data
--   Relies on GenericConfigurationManagerImpl<> classes to implement most API
-    functionality
--   Delegates low-level reading and writing of persistent values to ZephyrConfig
-    class
+-   Uses Zephyr platform implementation of ConfigurationManager interface.
 
 `include/platform/nrfconnect/ConnectivityManagerImpl.h`<br>`nrfconnect/ConnectivityManagerImpl.cpp`
 
@@ -30,30 +21,20 @@ this code will have parallels in any new adaptation.
     functionality
 -   Very much a work-in-progress in the nrfconnect branch
 
-`include/platform/nrfconnect/ThreadStackManagerImpl.h`<br>`nrfconnect/ThreadStackManagerImpl.cpp`
+`include/platform/nrfconnect/DeviceNetworkProvisioningDelegateImpl.h`<br>`nrfconnect/DeviceNetworkProvisioningDelegateImpl.cpp`
 
--   Concrete implementation of ThreadStackManager interface
--   Supports Thread stack initialization and core event loop processing
--   Relies on GenericThreadStackManagerImpl_OpenThread/POSIX<> classes to
-    implement most API functionaltiy
+-   Concrete implementation of DeviceNetworkProvisioningDelegate interface
+-   Provides high-level APIs for managing network provisioning
 
-`include/platform/nrfconnect/BLEManagerImpl.h`<br>`nrfconnect/BLEManagerImpl.cpp`
+`include/platform/nrfconnect/ThreadStackManagerImpl.h`
 
--   Concrete implementation of the BLEManager interface
--   Maps CHIP's BLE interface abstractions (BleLayer, BlePlatformDelegate,
-    BleApplicationDelegate) onto the platform's native BLE services
--   Implements chip-compatible BLE advertising.
+-   Uses Zephyr platform implementation of ThreadStackManager interface.
 
-`platform/nrfconnect/Entropy.cpp`
+`include/platform/nrfconnect/BLEManagerImpl.h`
 
--   Implements interface to platform entropy source
+-   Uses Zephyr platform implementation of BLEManager interface.
 
-`platform/nrfconnect/Logging.cpp`
+`include/platform/nrfconnect/CHIPDevicePlatformEvent.h`
 
--   Adaption of chip debug logging to platform logging facility.
-
-`platform/nrfconnect/ZephyrConfig.cpp`
-
--   Implements low-level read/write of persistent configuration values
--   Class API specifically designed to work in conjunction with the
-    GenericConfigurationManagerImpl<> class.
+-   Uses Zephyr platform definitions of platform-specific event types and data
+    for the chip Device Layer.
