+++ src/platform/nrfconnect/ThreadStackManagerImpl.h
@@ -17,75 +17,11 @@
 
 /**
  *    @file
- *          Provides an implementation of the ThreadStackManager object for
- *          nRF Connect SDK platforms.
+ *          Provides an implementation of the ThreadStackManager object
+ *          for nRF Connect SDK platforms, by including Zephyr platform
+ *          implementation.
  */
 
 #pragma once
 
-#include <platform/OpenThread/GenericThreadStackManagerImpl_OpenThread.h>
-#include <platform/Zephyr/GenericThreadStackManagerImpl_Zephyr.h>
-
-#include <openthread/thread.h>
-#include <platform/nrfconnect/BLEManagerImpl.h>
-
-namespace chip {
-namespace DeviceLayer {
-
-class ThreadStackManager;
-class ThreadStackManagerImpl;
-
-/**
- * Concrete implementation of the ThreadStackManager singleton object for nRF Connect platforms.
- */
-class ThreadStackManagerImpl final : public ThreadStackManager,
-                                     public Internal::GenericThreadStackManagerImpl_Zephyr<ThreadStackManagerImpl>
-{
-    // Allow the ThreadStackManager interface class to delegate method calls to
-    // the implementation methods provided by this class.
-    friend class ThreadStackManager;
-
-    // Allow the generic implementation base classes to call helper methods on
-    // this class.
-#ifndef DOXYGEN_SHOULD_SKIP_THIS
-    friend Internal::GenericThreadStackManagerImpl_OpenThread<ThreadStackManagerImpl>;
-    friend Internal::GenericThreadStackManagerImpl_Zephyr<ThreadStackManagerImpl>;
-#endif
-
-private:
-    // ===== Members for internal use by the following friends.
-
-    friend ThreadStackManager & ::chip::DeviceLayer::ThreadStackMgr(void);
-    friend ThreadStackManagerImpl & ::chip::DeviceLayer::ThreadStackMgrImpl(void);
-
-    static ThreadStackManagerImpl sInstance;
-
-    // ===== Private members for use by this class only.
-
-    ThreadStackManagerImpl() = default;
-};
-
-/**
- * Returns the public interface of the ThreadStackManager singleton object.
- *
- * chip applications should use this to access features of the ThreadStackManager object
- * that are common to all platforms.
- */
-inline ThreadStackManager & ThreadStackMgr(void)
-{
-    return ThreadStackManagerImpl::sInstance;
-}
-
-/**
- * Returns the platform-specific implementation of the ThreadStackManager singleton object.
- *
- * chip applications can use this to gain access to features of the ThreadStackManager
- * that are specific to nRF Connect platforms.
- */
-inline ThreadStackManagerImpl & ThreadStackMgrImpl(void)
-{
-    return ThreadStackManagerImpl::sInstance;
-}
-
-} // namespace DeviceLayer
-} // namespace chip
+#include <platform/Zephyr/ThreadStackManagerImpl.h>
